Magnetometer
The first sensor to be include is the magnetometer. The magnetometer used is built into the MPU9250 module. This magnetometer is used to act as a compass for a sense of direction.  Drivers are taken from kriswiner's git repo for this module https://github.com/kriswiner/MPU9250 and the magcal is based on a dicussions of this forum https://github.com/kriswiner/MPU6050/wiki/Simple-and-Effective-Magnetometer-Calibration